backend/requirements.txt
# Flask Framework
Flask
flask-cors

# Database
SQLAlchemy
Flask-SQLAlchemy

# Authentication
Flask-JWT-Extended
bcrypt

# LangChain & AI
langchain
langgraph
langchain-openai
langchain-community


# Vector Stores & Embeddings
chromadb
faiss-cpu
# sentence-transformers==2.3.1

# Document Processing
pypdf
python-docx
python-multipart

# Utilities
python-dotenv
marshmallow
requests
pydantic

# Web Search (Optional)
duckduckgo-search

# Development
watchdog
flask-restx
